REFACTOR: generateRandomNumber 모듈 추출
